timestamps are float, and int() won't convert that

+++ python-modules/cis_notifications/cis_notifications/event.py
@@ -79,13 +79,13 @@ class Event(object):
 
             # Check if what we had in secrets is still valid!
             # This includes 10s leeway for clock sync issues
-            if int(self.access_token_dict["exp"]) < time.time() - 10:
+            if float(self.access_token_dict["exp"]) < time.time() - 10:
                 logger.info("Access token has expired, refreshing")
                 authzero = self._get_authzero_client()
                 self.access_token_dict = authzero.exchange_for_access_token()
                 # Auth0 gives us the difference (expires_in) not a time stamp, so we need to calculate when the token
                 # expires. On failure to read expires_in, just make it expire in 60s as fallback
-                self.access_token_dict["exp"] = time.time() + int(self.access_token_dict.get("expires_in"), 60)
+                self.access_token_dict["exp"] = time.time() + float(self.access_token_dict.get("expires_in", 60.0))
                 self.secret_manager.secretmgr_store("az_access_token", self.access_token_dict)
             else:
                 logger.info("Re-using cached access token")
